# docs/swagger.yaml
basePath: /
definitions:
...
  types.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  types.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/types.APIError'
      success:
        type: boolean
    type: object
...
  types.CreateNotificationRequest:
    properties:
      bot_token:
        description: telegram
        type: string
      channel:
        description: 渠道,telegram,lark,feishu
        type: string
      chat_id:
        description: 聊天ID
        type: string
      name:
        description: 通用
        type: string
      secret:
        description: 签名验证时的密钥
        type: string
      webhook_url:
        description: lark feishu
        type: string
    required:
    - channel
    - name
    type: object
...
  types.DeleteNotificationRequest:
    properties:
      channel:
        description: 渠道,telegram,lark,feishu
        type: string
      name:
        description: 通用
        type: string
    required:
    - channel
    - name
    type: object
...
  types.FeishuConfig:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: ID
        type: integer
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 名称
        type: string
      secret:
        description: 签名验证时的密钥
        type: string
      updated_at:
        description: 更新时间
        type: string
      user_address:
        description: 用户地址
        type: string
      webhook_url:
        description: 网络钩子URL
        type: string
    type: object
...
  types.LarkConfig:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: ID
        type: integer
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 名称
        type: string
      secret:
        description: 签名验证时的密钥
        type: string
      updated_at:
        description: 更新时间
        type: string
      user_address:
        description: 用户地址
        type: string
      webhook_url:
        description: 网络钩子URL
        type: string
    type: object
...
  types.NotificationConfig:
    properties:
      bot_token:
        description: 可选字段，不同渠道用不同
        type: string
      channel:
        description: telegram / lark / feishu
        type: string
      chat_id:
        type: string
      created_at:
        type: string
      id:
        description: 通用
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      secret:
        type: string
      updated_at:
        type: string
      user_address:
        type: string
      webhook_url:
        type: string
    type: object
  types.NotificationConfigListResponse:
    properties:
      feishu_configs:
        items:
          $ref: '#/definitions/types.FeishuConfig'
        type: array
      lark_configs:
        items:
          $ref: '#/definitions/types.LarkConfig'
        type: array
      telegram_configs:
        items:
          $ref: '#/definitions/types.TelegramConfig'
        type: array
    type: object
...
  types.TelegramConfig:
    properties:
      bot_token:
        description: 机器人token
        type: string
      chat_id:
        description: 聊天ID
        type: string
      created_at:
        description: 创建时间
        type: string
      id:
        description: ID
        type: integer
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 名称
        type: string
      updated_at:
        description: 更新时间
        type: string
      user_address:
        description: 用户地址
        type: string
    type: object
...
  types.UpdateNotificationRequest:
    properties:
      bot_token:
        description: telegram
        type: string
      channel:
        description: 渠道,telegram,lark,feishu
        type: string
      chat_id:
        description: 聊天ID
        type: string
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 通用
        type: string
      secret:
        description: 签名验证时的密钥
        type: string
      webhook_url:
        description: lark feishu
        type: string
    required:
    - channel
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: TimeLocker Backend API
  title: TimeLocker Backend API
  version: "1.0"
paths:
...
  /api/v1/notifications/configs:
    post:
      consumes:
      - application/json
      description: 获取当前用户的所有通知渠道配置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.NotificationConfigListResponse'
              type: object
        "401":
          description: 未认证
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 获取所有通知配置
      tags:
      - Notification
  /api/v1/notifications/create:
    post:
      consumes:
      - application/json
      description: 为当前用户创建新的通知配置, 名字的空格会被自动去除, 防止攻击者通过空格来绕过名称验证
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.NotificationConfig'
              type: object
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "409":
          description: 配置名称已存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 创建通知配置
      tags:
      - Notification
  /api/v1/notifications/delete:
    post:
      consumes:
      - application/json
      description: 删除当前用户的通知配置, 名字的空格会被自动去除, 防止攻击者通过空格来绕过名称验证
      parameters:
      - description: 删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.DeleteNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: 配置不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 删除通知配置
      tags:
      - Notification
  /api/v1/notifications/update:
    post:
      consumes:
      - application/json
      description: 更新当前用户的通知配置, 如果不需要更新某个字段, 可以不传该字段, 但至少传一个字段
      parameters:
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: 配置不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 更新通知配置
      tags:
      - Notification
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"